refactor: rename remark-form-hoc
snake case is more conventional in this project
